added 10x less granularity when in idle mode
changed data type to float32, should have no loss of precision with integers up to 22 bits

+++ src/COMPILE_OPTS.h
@@ -6,6 +6,6 @@
 #define DAQ_DAEMON_COMPILE_OPTS_H
 #include <cstdint>
 
-typedef int32_t DATA_TYPE;
+typedef float DATA_TYPE;
 
 #endif //DAQ_DAEMON_COMPILE_OPTS_H
